refactor: 크루 랭커 score 필드를 double 에서 long 으로 변경

+++ src/main/java/revi1337/onsquad/crew_member/application/response/CrewRankedMemberResponse.java
@@ -9,7 +9,7 @@ public record CrewRankedMemberResponse(
         String nickname,
         String mbti,
         int rank,
-        double score,
+        long score,
         LocalDateTime lastActivityTime
 ) {
